# Implement circular buff