The purpose of this PR is to implement a circular buffer and then utilize that buffer instead of a slice for ngrams.

```bash
benchmark                old ns/op     new ns/op     delta
Benchmark_toNGrams-8     708           566           -20.01%
```